refactor: move then_webhook_notification to uc002 domain step file (aya0)
