+++ tests/bdd/steps/generic/then_media_buy.py
@@ -350,183 +350,6 @@ def then_slack_notification_sent(ctx: dict) -> None:
             )
 
 
-@then("the Buyer should be notified via webhook")
-def then_webhook_notification(ctx: dict) -> None:
-    """Assert buyer webhook notification dispatch prerequisites and payload correctness.
-
-    Production delivery path (src/core/context_manager.py::_send_push_notifications):
-      1. Query ObjectWorkflowMapping rows by step_id.
-      2. Query PushNotificationConfig (tenant_id, principal_id, is_active=True).
-      3. Read ``push_notification_config.url`` from step.request_data.
-      4. Build payload (media_buy_id, status, rejection_reason) and POST to the URL.
-
-    Hard assertions (all verified, all pass):
-      A. PushNotificationConfig row: url matches, is_active=True, principal_id matches.
-      C. ObjectWorkflowMapping exists linking step_id to the media buy.
-      D. Media buy + workflow step are in terminal status.
-      E. Notification payload content: media buy carries rejection status and
-         non-empty rejection_reason (the data the webhook would deliver).
-
-    Targeted xfail (harness gap -- only this check is xfailed):
-      B. step.request_data carries push_notification_config URL -- required for
-         _send_push_notifications to actually POST. The BDD reject path uses
-         repository methods that bypass the admin flow which populates this field.
-         FIXME(salesagent-9vgz.1): Wire through the production admin approve/reject
-         flow, then remove the xfail.
-    """
-    import pytest
-    from sqlalchemy import select
-
-    from src.core.database.models import ObjectWorkflowMapping, PushNotificationConfig
-    from src.core.database.repositories.media_buy import MediaBuyRepository
-    from src.core.database.repositories.workflow import WorkflowRepository
-
-    # --- Extract media_buy_id and tenant ---
-    resp = ctx.get("response")
-    existing_mb = ctx.get("existing_media_buy")
-    assert resp is not None or existing_mb is not None, (
-        "No response or existing media buy in ctx — nothing to notify the Buyer about"
-    )
-
-    media_buy_id = None
-    if resp is not None:
-        media_buy_id = _get_response_field(resp, "media_buy_id")
-    elif existing_mb is not None:
-        media_buy_id = getattr(existing_mb, "media_buy_id", None)
-    assert media_buy_id, "No media_buy_id — cannot verify notification"
-
-    tenant = ctx.get("tenant")
-    assert tenant is not None, "No tenant in ctx — cannot verify notification scoping"
-    tenant_id = getattr(tenant, "tenant_id", None) or (tenant.get("tenant_id") if isinstance(tenant, dict) else None)
-
-    # --- Check push_notification_config was registered by Given step ---
-    push_config = ctx.get("push_notification_config")
-    assert push_config is not None, (
-        "No push_notification_config in ctx — scenario must include a Given step "
-        "that sets ctx['push_notification_config'] with the expected webhook URL."
-    )
-    expected_url = push_config.get("url") if isinstance(push_config, dict) else None
-    assert expected_url, "push_notification_config has no 'url' — cannot verify webhook destination"
-
-    # --- A. PushNotificationConfig row: exact url match, active, correct principal ---
-    principal = ctx.get("principal")
-    expected_principal_id = (
-        getattr(principal, "principal_id", None)
-        if principal is not None
-        else (principal.get("principal_id") if isinstance(principal, dict) else None)
-    )
-    with _db_session(ctx) as session:
-        configs = (
-            session.scalars(select(PushNotificationConfig).filter_by(tenant_id=tenant_id, url=expected_url)).all() or []
-        )
-        stored_urls = [c.url for c in configs]
-        assert expected_url in stored_urls, (
-            f"Expected webhook URL '{expected_url}' not found in PushNotificationConfig "
-            f"for tenant {tenant_id}. Stored URLs: {stored_urls}. "
-            "Dispatcher will not find the webhook destination."
-        )
-        assert any(c.is_active for c in configs), (
-            f"PushNotificationConfig rows for url={expected_url} exist but none have is_active=True "
-            f"(found: {[(c.id, c.is_active) for c in configs]}) — "
-            "_send_push_notifications filters by is_active=True and will skip them"
-        )
-        if expected_principal_id:
-            active_principal_ids = [c.principal_id for c in configs if c.is_active]
-            assert expected_principal_id in active_principal_ids, (
-                f"No active PushNotificationConfig for principal_id={expected_principal_id}; "
-                f"active rows belong to principals: {active_principal_ids}. "
-                "Dispatcher filters by principal_id — the webhook would be addressed to the wrong buyer."
-            )
-
-    # --- D. Media buy in terminal status (status-change trigger has fired) ---
-    with _db_session(ctx) as session:
-        mb_repo = MediaBuyRepository(session, tenant_id)
-        mb = mb_repo.get_by_id(str(media_buy_id))
-        assert mb is not None, f"Media buy {media_buy_id} not found — cannot verify status change"
-        terminal_statuses = {"rejected", "approved", "active", "completed", "cancelled"}
-        assert mb.status in terminal_statuses, (
-            f"Media buy {media_buy_id} has status '{mb.status}' — expected a terminal "
-            f"status ({terminal_statuses}) proving the status-change event that "
-            "triggers webhook delivery has occurred"
-        )
-
-    # --- E. Notification payload content ---
-    # _send_push_notifications builds the webhook payload from the media buy's
-    # current state. Verify the media buy carries the data the buyer expects:
-    # for rejection, the payload must include a non-empty rejection_reason.
-    with _db_session(ctx) as session:
-        mb_repo = MediaBuyRepository(session, tenant_id)
-        mb = mb_repo.get_by_id(str(media_buy_id))
-        assert mb is not None, f"Media buy {media_buy_id} disappeared between checks"
-        if mb.status == "rejected":
-            assert mb.rejection_reason is not None and mb.rejection_reason.strip() != "", (
-                f"Media buy {media_buy_id} has status 'rejected' but rejection_reason is "
-                f"'{mb.rejection_reason}' — webhook payload would lack the rejection reason, "
-                "violating the Buyer notification contract (POST-S12)"
-            )
-
-    # --- C. Workflow step + mapping ---
-    with _db_session(ctx) as session:
-        wf_repo = WorkflowRepository(session, tenant_id)
-        mapping = wf_repo.get_latest_mapping_for_object("media_buy", str(media_buy_id))
-        assert mapping is not None, (
-            f"No workflow mapping for media_buy {media_buy_id} — "
-            "_send_push_notifications iterates ObjectWorkflowMapping; with none, nothing is dispatched"
-        )
-        # Mapping must point at the right object (dispatcher uses object_type + object_id in payload).
-        assert mapping.object_type == "media_buy", (
-            f"Expected mapping.object_type='media_buy', got '{mapping.object_type}' — "
-            "dispatcher would build the wrong payload kind"
-        )
-        assert str(mapping.object_id) == str(media_buy_id), (
-            f"Mapping object_id='{mapping.object_id}' != media_buy_id='{media_buy_id}' — "
-            "dispatcher would address a different object"
-        )
-
-        step = wf_repo.get_step_by_id(mapping.step_id)
-        assert step is not None, (
-            f"Workflow step {mapping.step_id} not found — context_manager cannot dispatch without a step record"
-        )
-        terminal_step_statuses = {"rejected", "completed", "approved", "failed"}
-        assert step.status in terminal_step_statuses, (
-            f"Workflow step {mapping.step_id} has status '{step.status}' — "
-            f"expected one of {terminal_step_statuses} so the status-change event fires dispatch"
-        )
-
-        # Cross-check the mapping links back to this step (dispatcher reads step.request_data).
-        mappings_for_step = session.scalars(select(ObjectWorkflowMapping).filter_by(step_id=step.step_id)).all()
-        assert mappings_for_step, (
-            f"No mappings discoverable by step_id={step.step_id} — "
-            "_send_push_notifications queries ObjectWorkflowMapping by step_id and would find nothing"
-        )
-
-        # --- B. step.request_data must carry push_notification_config with the buyer URL ---
-        # _send_push_notifications reads step.request_data['push_notification_config']['url']
-        # to determine the webhook destination. Without it, dispatch logs
-        # 'No push notification URL present' and skips the POST entirely.
-        #
-        # SPEC-PRODUCTION GAP: the repository-driven reject path does NOT populate
-        # step.request_data with push_notification_config because it bypasses the
-        # Flask admin flow that writes the original request payload onto the step.
-        # FIXME(salesagent-9vgz.1): wire through the production admin approve/reject
-        # flow which populates request_data, then remove this xfail.
-        req_data = step.request_data or {}
-        step_push_cfg = req_data.get("push_notification_config") if isinstance(req_data, dict) else None
-        if not isinstance(step_push_cfg, dict) or step_push_cfg.get("url") != expected_url:
-            pytest.xfail(
-                "SPEC-PRODUCTION GAP: step.request_data does not carry "
-                "push_notification_config with the buyer's URL — "
-                "_send_push_notifications would skip dispatch. "
-                "FIXME(salesagent-9vgz.1): wire through the admin flow."
-            )
-
-        # Happy path (reached when harness wires the full admin flow):
-        assert step_push_cfg["url"] == expected_url, (
-            f"step.request_data push_notification_config URL mismatch: "
-            f"expected '{expected_url}', got '{step_push_cfg.get('url')}'"
-        )
-
-
 # ═══════════════════════════════════════════════════════════════════════
 # Persistence assertions
 # ═══════════════════════════════════════════════════════════════════════
